requirements.txt
discord==2.1.0
dpytest==0.6.1
python-dotenv==0.21.1